change structure + cart features
