+++ src/Pages/Cart/Cart.css
@@ -1,0 +1,12 @@
+.cart {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+}
+
+.cart-content {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+}
