+++ src/Pages/Cart/Cart.jsx
@@ -1,8 +1,11 @@
 import "./Cart.css"
+import CartViewer from "../../components/ItemComponents/CartViewer/CartViewer"
 function Cart() {
     return(
-        <div>
-            I am a cart
+        <div className="cart">
+            <div className="cart-content">
+                <CartViewer/>
+            </div>
         </div>
     )
 }
