--- src/components/Dropdown/Dropdown.css
@@ -1,5 +1,0 @@
-.dropdown {
-    position: relative;
-    margin-left: 5px;
-    margin-right: 5px;
-}
