+++ src/components/DropdownComponents/Dropdown/Dropdown.css
@@ -1,0 +1,5 @@
+.dropdown {
+    position: relative;
+    margin-left: 5px;
+    margin-right: 5px;
+}
