+++ src/components/DropdownComponents/DropdownContent/DropdownContent.css
@@ -1,0 +1,26 @@
+.dropdown-content {
+    position: absolute;
+    background-color: white;
+    min-width: 100%;
+    width: max-content;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    max-height: 40vh;
+    overflow-y: scroll;
+    scrollbar-width: none;
+    -ms-overflow-style: none;
+    border: 2px solid hsla(0, 0%, 60%, 0.3);
+    border-radius: 5px;
+    opacity: 0;
+    color: #667085;
+    transform: translateY(-10%);
+    transition: transform 100ms ease-in-out, opacity 100ms ease-in-out;
+    pointer-events: none;
+}
+
+.dropdown-content-open {
+    opacity: 1;
+    transform: translateY(0);
+    pointer-events: all;
+}
