+++ src/components/HeaderComponents/Header/Header.css
@@ -1,13 +1,14 @@
 .header {
+    position: sticky;
+    top: 0;
     display: inline-flex;
     width: 100%;
     height: 60px;
-    background-color: white;
-    border: 1px solid black;
     justify-content: center;
     background-color: aqua;
-    margin-bottom: 10px;
     flex-basis: 0;
+    z-index: 1000;
+    padding-bottom: 3px;
 }
 
 .brand-signature {
