+++ src/components/HeaderComponents/NavigationBar/NavigationBar.css
@@ -1,0 +1,23 @@
+.nav-bar {
+    width: 100%;
+    height: 5vh;
+    display: grid;
+    grid-template-columns: repeat(3, 10%);
+    background-color: aqua;
+    justify-content: center;
+    align-content: center;
+    text-decoration: none;
+    color: black;
+    margin-bottom: 10px;
+    border: 1px solid hsl(0, 0%, 80%);
+}
+
+.nav-option {
+    text-align: center;
+    text-decoration: none;
+    color: black;
+    font-size: 1.2rem;
+    font-weight: bold;
+    font-family: MS Sans Serif4, Geneva, sans-serif;
+    cursor: pointer;
+}
