+++ src/components/ItemComponents/CartViewer/CartViewer.css
@@ -1,0 +1,10 @@
+.cart-viewer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: fit-content;
+    max-width: 70%;
+    padding: 20px;
+    border: 1cap solid hsl(0, 0%, 0%, 0.1);
+    border-radius: 5px;
+}
