--- src/components/ItemViewer/ItemViewer.css
@@ -1,36 +1,0 @@
-.item-viewer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: fit-content;
-    max-width: 70%;
-    padding: 20px;
-    border: 1cap solid hsl(0, 0%, 0%, 0.1);
-    border-radius: 5px;
-}
-
-.dropdown-container {
-    position: sticky;
-    top: 0;
-    background-color: white;
-    width: 100%;
-    display: flex;
-    align-items: flex-start;
-}
-
-.product-container {
-    display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 10px;
-    width: fit-content;
-    margin-top: 10px;
-    margin-bottom: 10px;
-}
-
-.item-viewer-small {
-    max-width: 90%;
-}
-
-.product-container-small {
-    grid-template-columns: repeat(2,1fr);
-}
