+++ src/customHook/formatPrice.jsx
@@ -1,0 +1,3 @@
+export default function formatPrice(price) {
+    return price.toLocaleString('de-DE');
+}
